practices/pr_cache/requirements.txt
redis>=4.6.0
psycopg2-binary>=2.9.6
